# docs/api-openapi.yaml
openapi: 3.1.0
info:
  title: DischargeGuide API
  version: 0.1.0-placeholder
  description: Safe development contract. Clinical integrations are not live yet.
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
paths:
...
  /upload:
    post:
      summary: Upload discharge instructions and start processing
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [document]
              properties:
                document:
                  type: string
                  format: binary
      responses:
        "202":
          description: Processing started
          content:
            application/json:
              schema:
                type: object
                required: [documentId, status, processUrl, isPlaceholder]
                properties:
                  documentId: { type: string, format: uuid }
                  status: { const: processing }
                  processUrl: { type: string }
                  isPlaceholder: { type: boolean }
        "415": { description: Unsupported file type }
...
  /medications:
    get:
      summary: List medications for a recovery plan
      parameters:
        - in: query
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Medication list }
...
  /appointments:
    get:
      summary: List upcoming appointments
      parameters:
        - in: query
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Appointment list }
  /appointments/{appointmentId}/calendar:
    post:
      summary: Download an appointment as an ICS calendar event
      parameters:
        - in: path
          name: appointmentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Calendar file
          content:
            text/calendar: { schema: { type: string } }
  /ask:
    post:
      summary: Ask a document-grounded question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question, documentId]
              properties:
                question: { type: string }
                documentId: { type: string, format: uuid }
      responses:
        "200": { description: Grounded answer or explicit unavailable response }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT